requirements.txt
click>=8.0.0
rich>=13.0.0
pyyaml>=6.0
mlx>=0.15.0
mlx-lm>=0.15.0
transformers>=4.30.0
tokenizers>=0.13.0
huggingface-hub>=0.16.0
pandas>=1.5.0
numpy>=1.21.0
tqdm>=4.65.0
psutil>=5.9.0